requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
django-embed-video==1.4.7
idna==3.4
Pillow==9.2.0
requests==2.28.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.12
